Q&A게시판 글목록 글등록 완료

+++ ict/src/main/java/com/newdeal/ict/Service/QaBoardService.java
@@ -1,6 +1,5 @@
 package com.newdeal.ict.Service;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 import javax.servlet.http.HttpSession;
@@ -9,7 +8,7 @@ import com.newdeal.ict.Vo.QaBoardVo;
 
 public interface QaBoardService {
 	public int create(QaBoardVo vo) throws Exception;//글쓰기
-	public QaBoardVo read(BigDecimal qaNum) throws Exception;
+	public QaBoardVo read(int qaNum) throws Exception;
 	public void update(QaBoardVo vo) throws Exception;
 	public void delete(int qaNum) throws Exception;//글삭제
 	public List<QaBoardVo> listAll(
